React Router v6 withRouter replacement and implementing hooks with Profile form
Still trying to fix CreateProfile.js and adapting it for use with react router 6. Created ProfileForm.js to replace the old script but ran into issues. It does not render the create profile form as it used to in the old code despite tweaks to the onChange, onSubmit, export values, useEffect() and the navigate/useNavigate (adapted) fucntionality.