FINALLY GOT IT TO WORK WITH REACT ROUTER 6 PROTOCOLS
Tweaked the code to work now. Had to debug almost all the profile-related files to get it to function properly. Hooks are making things easier but this was made a lot more complex than it needed to be due to how much the structure changed since router v5. Profiles can now be added successfully added to logged in users. Private route protocol has also been debugged and works now.