Developed the Edit profile page and route
completed the edit profile page as per both, old and new convention (i.e. react router 5 has EditProfile.js and 6 still only relies on ProfileForm.js). It is for some reason not loading the profile data already in the DB. We get hit with an error 500 so I am re-evaluating and refactoring the profile.js backend code to check reasons for incompatibility/errors. FIX ME: PROFILE DATA NOT LOADING FROM BACKEND; THE DASHBOARD KEEPS ON ASKING US TO CREATE A PROFILE FOR A USER WHICH ALREADY HAS ONE