Developed the addEducation front end
Completed developing the addEducation page and field. Tested to be working. The fields register successfully and all the form data gets carried over properly. The issue with going back, using 'history'(?) and having to create a new profile on every login, when it doesn't load the old still persists though. FIX ME: GOING BACK GOES TO A BLANK DASHBOARD IF DONE VIA BUTTONS, THE PROFILE NEEDS TO BE CREATED EVERY TIME SINCE IT DOESN'T PICK UP AN ALREADY ESTABLISHED PROFILE FROM THE DB FOR SOME REASON.