Added the delete functionality to the delete buttons
Developed the delete functionality and connected it to the respective buttons for education, experience and account values. Also developed the functionality to delete the posts associated with the account on account deletion at the back end. FIX ME: LOGIN DASHBOARD DOESN'T LOAD PROFILE AND THE BACK BUTTON COMPONENTS STILL DON'T WORK.