Heavily refactored code (in sheer 'files edited' terms). Resolved the FIX ME Issues.
Refactored a lot of files all the way to the backend (where the fetch profile on login issue was arising from. It was missing a return.json(profile) statement in the try block. The dashboard profileforms, addXp and addEdu back issue was being caused by an incorrect link passed as the argument.