Developed the developers section of the website.
Developed the page to render all user profiles on developers on the developers page of website. It is breaking down at the profileItem component section though. For some reason, it seems to be picking up the profile as null (I suspect it's trying to access the current profile data instead of picking up the profiles which are a subsection of the profiles array. FIX ME: FIX PROFILE-PROFILEITEM COMPONENTS.